Finalize Dev Null for Release
