--- airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/check/DestinationCheckOperation.kt
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.cdk.check
-
-import io.airbyte.cdk.command.DestinationConfiguration
-
-/**
- * A check operation that is run before the destination is used.
- *
- * * Implementors must provide a [check] method that validates the configuration.
- * * Implementors may provide a [cleanup] method that is run after the check is complete.
- * * [check] should throw an exception if the configuration is invalid.
- * * [cleanup] should not throw exceptions.
- * * Implementors should not perform any side effects in the constructor.
- * * Implementors should not throw exceptions in the constructor.
- * * Implementors should not inject configuration; only use the config passed in [check].
- */
-interface DestinationCheckOperation<C : DestinationConfiguration> {
-    fun check(config: C)
-    fun cleanup() {}
-}
