+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/command/DestinationConfiguration.kt
@@ -7,13 +7,13 @@ package io.airbyte.cdk.command
 import io.micronaut.context.annotation.Factory
 import jakarta.inject.Singleton
 
-interface DestinationConfiguration : Configuration {
-    val recordBatchSizeBytes: Long
-    val firstStageTmpFilePrefix: String
+abstract class DestinationConfiguration : Configuration {
+    val recordBatchSizeBytes: Long = 200L * 1024L * 1024L
+    val firstStageTmpFilePrefix: String = "airbyte-cdk-load-staged-raw-records"
 
     /** Memory queue settings */
-    val maxMessageQueueMemoryUsageRatio: Double // as fraction of available memory
-    val estimatedRecordMemoryOverheadRatio: Double // 0 => No overhead, 1.0 => 2x overhead
+    val maxMessageQueueMemoryUsageRatio: Double = 0.2 // 0 => No limit, 1.0 => 100% of JVM heap
+    val estimatedRecordMemoryOverheadRatio: Double = 0.1 // 0 => No overhead, 1.0 => 100% overhead
 
     /**
      * Micronaut factory which glues [ConfigurationJsonObjectSupplier] and
@@ -28,12 +28,5 @@ interface DestinationConfiguration : Configuration {
         ): DestinationConfiguration {
             return factory.make(pojoSupplier.get())
         }
-    }
-
-    companion object {
-        const val DEFAULT_RECORD_BATCH_SIZE_BYTES: Long = 200L * 1024L * 1024L
-        const val DEFAULT_FIRST_STAGE_TMP_FILE_PREFIX = "airbyte-cdk-load-staged-raw-records"
-        const val DEFAULT_MAX_MESSAGE_QUEUE_MEMORY_USAGE_RATIO: Double = 0.2
-        const val DEFAULT_ESTIMATED_RECORD_MEMORY_OVERHEAD_RATIO: Double = 0.1
     }
 }
