+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/task/SetupTask.kt
@@ -4,18 +4,18 @@
 
 package io.airbyte.cdk.task
 
-import io.airbyte.cdk.write.DestinationWriteOperation
+import io.airbyte.cdk.write.DestinationWriter
 import io.micronaut.context.annotation.Secondary
 import jakarta.inject.Singleton
 
 /**
- * Wraps @[DestinationWriteOperation.setup] and starts the open stream tasks.
+ * Wraps @[DestinationWriter.setup] and starts the open stream tasks.
  *
  * TODO: This should call something like "TaskLauncher.setupComplete" and let it decide what to do
  * next.
  */
 class SetupTask(
-    private val destination: DestinationWriteOperation,
+    private val destination: DestinationWriter,
     private val taskLauncher: DestinationTaskLauncher
 ) : Task {
     override suspend fun execute() {
@@ -27,7 +27,7 @@ class SetupTask(
 @Singleton
 @Secondary
 class SetupTaskFactory(
-    private val destination: DestinationWriteOperation,
+    private val destination: DestinationWriter,
 ) {
     fun make(taskLauncher: DestinationTaskLauncher): SetupTask {
         return SetupTask(destination, taskLauncher)
