+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/task/TeardownTask.kt
@@ -5,19 +5,19 @@
 package io.airbyte.cdk.task
 
 import io.airbyte.cdk.state.StreamsManager
-import io.airbyte.cdk.write.DestinationWriteOperation
+import io.airbyte.cdk.write.DestinationWriter
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.micronaut.context.annotation.Secondary
 import jakarta.inject.Singleton
 import java.util.concurrent.atomic.AtomicBoolean
 
 /**
- * Wraps @[DestinationWriteOperation.teardown] and stops the task launcher.
+ * Wraps @[DestinationWriter.teardown] and stops the task launcher.
  *
  * TODO: Report teardown-complete and let the task launcher decide what to do next.
  */
 class TeardownTask(
-    private val destination: DestinationWriteOperation,
+    private val destination: DestinationWriter,
     private val streamsManager: StreamsManager,
     private val taskLauncher: DestinationTaskLauncher
 ) : Task {
@@ -44,7 +44,7 @@ class TeardownTask(
 @Singleton
 @Secondary
 class TeardownTaskFactory(
-    private val destination: DestinationWriteOperation,
+    private val destination: DestinationWriter,
     private val streamsManager: StreamsManager,
 ) {
     fun make(taskLauncher: DestinationTaskLauncher): TeardownTask {
