+++ airbyte-integrations/connectors/destination-dev-null/metadata.yaml
@@ -1,8 +1,8 @@
 data:
-  connectorSubtype: file
+  connectorSubtype: unknown
   connectorType: destination
   definitionId: a7bcc9d8-13b3-4e49-b80d-d020b90045e3
-  dockerImageTag: 0.3.3
+  dockerImageTag: 0.7.0
   dockerRepository: airbyte/destination-dev-null
   githubIssueLabel: destination-dev-null
   icon: airbyte.svg
@@ -12,7 +12,7 @@ data:
     cloud:
       enabled: true
     oss:
-      enabled: false
+      enabled: true
   releaseStage: alpha
   documentationUrl: https://docs.airbyte.com/integrations/destinations/e2e-test
   tags:
