+++ airbyte-integrations/connectors/destination-dev-null/src/test/resources/test-catalog.json
