+++ airbyte-integrations/connectors/destination-dev-null/src/test/resources/test-data.jsonl
