+++ airbyte-integrations/connectors/destination-dev-null/test_configs/logging.json
