--- airbyte-integrations/connectors/destination-e2e-test/build.gradle
@@ -1,43 +1,0 @@
-plugins {
-    id 'application'
-    id 'airbyte-java-connector'
-}
-
-airbyteJavaConnector {
-    cdkVersionRequired = '0.20.6'
-    features = ['db-destinations']
-    useLocalCdk = false
-}
-
-application {
-    mainClass = 'io.airbyte.integrations.destination.e2e_test.E2EDestination'
-    applicationDefaultJvmArgs = ['-XX:+ExitOnOutOfMemoryError', '-XX:MaxRAMPercentage=75.0']
-
-    // Uncomment and replace to run locally
-    //applicationDefaultJvmArgs = ['-XX:+ExitOnOutOfMemoryError', '-XX:MaxRAMPercentage=75.0', '--add-opens', 'java.base/java.lang=ALL-UNNAMED']
-}
-
-// Uncomment to run locally
-//run {
-//    standardInput = System.in
-//}
-
-dependencies {
-    def vBulkCDK = '0.9'
-    def vMicronaut = '4.3.13'
-
-    ksp "io.micronaut:micronaut-inject-kotlin:${vMicronaut}"
-
-    //implementation "io.airbyte.bulk-cdk:bulk-cdk-core-base:${vBulkCDK}"
-    implementation project(":airbyte-cdk:bulk:core:bulk-cdk-core-base")
-    //implementation "io.airbyte.bulk-cdk:bulk-cdk-core-load:${vBulkCDK}"
-    implementation project(":airbyte-cdk:bulk:core:bulk-cdk-core-load")
-
-    kspTest "io.micronaut:micronaut-inject-kotlin:${vMicronaut}"
-
-    testImplementation platform('org.testcontainers:testcontainers-bom:1.19.8')
-    testImplementation project(":airbyte-cdk:bulk:core:bulk-cdk-core-load")
-    testImplementation testFixtures(project(":airbyte-cdk:bulk:core:bulk-cdk-core-load"))
-
-    // testImplementation testFixtures("io.airbyte.bulk-cdk:bulk-cdk-core-load:${vBulkCDK}"
-}
