--- airbyte-integrations/connectors/destination-e2e-test/src/main/kotlin/io/airbyte/integrations/destination/e2e_test/E2EDestination.kt
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.integrations.destination.e2e_test
-
-import io.airbyte.cdk.AirbyteDestinationRunner
-
-class E2EDestination {
-    companion object {
-        @JvmStatic
-        fun main(args: Array<String>) {
-            AirbyteDestinationRunner.run(*args)
-        }
-    }
-}
