--- airbyte-integrations/connectors/destination-e2e-test/src/main/kotlin/io/airbyte/integrations/destination/e2e_test/E2EDestinationSpecification.kt
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.integrations.destination.e2e_test
-
-import io.airbyte.cdk.spec.DestinationSpecification
-import io.airbyte.protocol.models.v0.DestinationSyncMode
-import jakarta.inject.Singleton
-
-@Singleton
-class E2EDestinationSpecification : DestinationSpecification {
-    override val supportedSyncModes =
-        listOf(
-            DestinationSyncMode.OVERWRITE,
-            DestinationSyncMode.APPEND,
-            DestinationSyncMode.APPEND_DEDUP,
-        )
-    override val supportsIncremental = true
-}
