--- airbyte-integrations/connectors/destination-e2e-test/src/test-integration/java/io/airbyte/integrations/destination/e2e_test/E2eCheckIntegrationTest.kt
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.integrations.destination.e2e_test
-
-import io.airbyte.cdk.test.check.CheckIntegrationTest
-
-class E2eCheckIntegrationTest :
-    CheckIntegrationTest<E2EDestinationConfigurationJsonObject>(
-        E2EDestinationConfigurationJsonObject::class.java,
-        successConfigFilenames = listOf(E2eTestUtils.LOGGING_CONFIG_PATH),
-        failConfigFilenamesAndFailureReasons = mapOf(),
-    )
