--- airbyte-integrations/connectors/destination-e2e-test/src/test-integration/java/io/airbyte/integrations/destination/e2e_test/E2eDestinationDataDumper.kt
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.integrations.destination.e2e_test
-
-import io.airbyte.cdk.test.util.DestinationDataDumper
-import io.airbyte.cdk.test.util.OutputRecord
-
-object E2eDestinationDataDumper : DestinationDataDumper {
-    override fun dumpRecords(streamName: String, streamNamespace: String?): List<OutputRecord> {
-        // E2e destination doesn't actually write records, so we shouldn't even
-        // have tests that try to read back the records
-        throw NotImplementedError()
-    }
-}
