--- airbyte-integrations/connectors/destination-e2e-test/src/test/resources/test-config.json
@@ -1,9 +1,0 @@
-{
-  "test_destination": {
-    "test_destination_type": "LOGGING",
-    "logging_config": {
-      "logging_type": "FirstN",
-      "max_entry_count": 100
-    }
-  }
-}
