--- test-catalog.json
@@ -1,27 +1,0 @@
-{
-  "streams": [
-    {
-      "stream": {
-        "name": "users",
-        "json_schema": {
-          "type": "object",
-          "properties": {
-            "id": {
-              "type": "integer"
-            },
-            "name": {
-              "type": "string"
-            }
-          }
-        }
-      },
-      "sync_mode": "full_refresh",
-      "cursor_field": [],
-      "destination_sync_mode": "overwrite",
-      "primary_key": [],
-      "generation_id": 10,
-      "minimum_generation_id": 10,
-      "sync_id": 101
-    }
-  ]
-}
