--- test-config-bad.json
@@ -1,9 +1,0 @@
-{
-  "test_destination": {
-    "test_destination_type": "LOGGING",
-    "logging_config": {
-      "logging_type": "FirstM",
-      "max_entry_count": 100
-    }
-  }
-}
