--- test-config-silent.json
@@ -1,5 +1,0 @@
-{
-  "test_destination": {
-    "test_destination_type": "SILENT"
-  }
-}
