--- test-data.jsonl
@@ -1,5 +1,0 @@
-{ "type": "RECORD", "record": { "emitted_at": 0, "stream": "users", "data": { "id": 1, "name": "Alice" } } }
-{ "type": "RECORD", "record": { "emitted_at": 1, "stream": "users", "data": { "id": 2, "name": "Bob" } } }
-{ "type": "RECORD", "record": { "emitted_at": 2, "stream": "users", "data": { "id": 3, "name": "Some Other Guy" } } }
-{ "type": "STATE", "state": { "ID": "pointless", "type": "STREAM", "stream": { "stream_descriptor": {  "name": "users" }, "stream_state": { "foo": "bar" } }, "sourceStats": { "recordCount": 3 } } }
-{ "type": "TRACE", "trace": { "emitted_at": 3, "type": "STREAM_STATUS", "stream_status": { "stream_descriptor": {  "name": "users" }, "status": "COMPLETE" } } }
